requirements.txt
botocore
cohere
voyageai
anthropic
elasticsearch
polars
numpy
stemmer
python-dotenv

llama_index
llama_index_core

llama-index-embeddings-huggingface
llama-index-llms-openrouter
llama-index-vector-stores-qdrant
llama-index-embeddings-nvidia

llama-index-llms-anthropic
llama-index-postprocessor-cohere-rerank
llama-index-retrievers-bm25

fastembed